feat: add types to burger-constructor component
